API_ASSIGN/requirements.txt
asgiref==3.7.2
Django==4.1.11
django-cors-headers==4.2.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.1
Pillow==10.0.0
pymongo==3.12.3
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.2.4
typing_extensions==4.7.1
uritemplate==4.1.1
